move config classes out of specific sub-package
